# Swagger/ucloud-1.0.0-oas3-swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0-oas3
  title: uCloud API collection
  description: uCloud is a collection of image APIs.
  contact:
    email: help@bps-eng.co.ir
  license:
    name: MIT
    url: 'http://www.opensource.org/licenses/MIT'
security:
  - implicit:
      - 'read:face'
      - 'write:face'
...
paths:
...
  /matchers:
    get:
      tags:
        - matching
      summary: List matchers
      description: Get the list of all matchers.
      operationId: matcherList
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Matcher'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Matcher'
      security:
        - implicit:
            - 'read:face'
    post:
      tags:
        - matching
      summary: Create matcher
      description: >-
        Create a new matcher which can be used to find similar models in input
        images.
      operationId: matcherNew
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Matcher'
          application/xml:
            schema:
              $ref: '#/components/schemas/Matcher'
        description: Matcher for finding models
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '201':
          description: successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - implicit:
            - 'write:face'
            - 'read:face'
  '/matchers/{id}':
    get:
      tags:
        - matching
      summary: View matcher
      description: View a matcher by id.
      operationId: matcherView
      parameters:
        - name: id
          in: path
          description: Matcher id to query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Matcher'
            application/xml:
              schema:
                $ref: '#/components/schemas/Matcher'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - implicit:
            - 'write:face'
            - 'read:face'
    post:
      tags:
        - matching
      summary: Update matcher with form data
      description: Update a matcher with form data.
      operationId: matcherUpdateForm
      parameters:
        - name: id
          in: path
          description: Matcher id to query
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  description: Updated matcher data
                  type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - implicit:
            - 'write:face'
            - 'read:face'
    put:
      tags:
        - matching
      summary: Update matcher
      description: Update a matcher.
      operationId: matcherUpdate
      parameters:
        - name: id
          in: path
          description: Matcher id to query
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Matcher'
          application/xml:
            schema:
              $ref: '#/components/schemas/Matcher'
        description: Updated matcher data
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '405':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - implicit:
            - 'write:face'
            - 'read:face'
    delete:
      tags:
        - matching
      summary: Delete a matcher
      description: Delete a matcher by id.
      operationId: matcherRemove
      parameters:
        - name: id
          in: path
          description: Matcher id to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - implicit:
            - 'write:face'
            - 'read:face'
  '/matchers/{id}/find':
    post:
      tags:
        - matching
      summary: Find model
      description: Find model instances in an image by the matcher.
      operationId: matcherFind
      parameters:
        - name: id
          in: path
          description: Matcher id to query
          required: true
          schema:
            type: string
        - name: min_score
          in: query
          description: Minimum match score
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 1
        - name: matches_count
          in: query
          description: >-
            Maximum count of matches to return; 0 means all matches should be
            returned.
          schema:
            type: integer
            minimum: 0
            default: 1
        - name: max_overlap
          in: query
          description: Maximum overlap of the model instances.
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 1
        - name: subpixel
          in: query
          description: Specify model localization accuracy.
          schema:
            type: string
            enum:
              - none
              - interpolation
              - least_squares
              - least_squares_high
              - least_squares_very_high
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Image'
          application/xml:
            schema:
              $ref: '#/components/schemas/Image'
        description: Image data for matching
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelInstance'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelInstance'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - implicit:
            - 'read:face'
...
servers:
  - url: 'https://virtserver.swaggerhub.com/sa-mustafa/ucloud/v1'
  - url: 'http://virtserver.swaggerhub.com/sa-mustafa/ucloud/v1'
components:
...
  securitySchemes:
    implicit:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
          scopes:
            'write:face': allows modifying resources
            'read:face': allows reading resources
    api_key:
      type: apiKey
      name: api_key
      in: header
  schemas:
...
    ModelInstance:
      type: object
      properties:
        angles:
          description: >-
            rotation angles of found instances of the model; applies to all
            except deformable/descriptor/perspective matchers.
          type: array
          items:
            type: number
            format: float
        cols:
          description: >-
            rows of found instances of the model; applies to all except
            descriptor/perspective matchers.
          type: array
          items:
            type: number
            format: float
        rows:
          description: >-
            columns of found instances of the model; applies to all except
            descriptor/perspective matchers.
          type: array
          items:
            type: number
            format: float
        scales:
          description: scales of found instances; applies to isotropic matcher only.
          type: array
          items:
            type: number
            format: float
        scales_row:
          description: >-
            scales of found instances in row direction; applies to anisotropic
            matcher only.
          type: array
          items:
            type: number
            format: float
        scales_col:
          description: >-
            scales of found instances in column direction; applies to
            anisotropic matcher only.
          type: array
          items:
            type: number
            format: float
        transforms:
          description: >-
            homography transforms between found instances and the model; applies
            to descriptor & perspective matchers only.
          type: array
          items:
            type: number
            format: float
        scores:
          description: matching scores of found instances.
          type: array
          items:
            type: number
            format: float
      xml:
        name: ModelInstance
...
    Image:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tag:
          type: string
        source:
          type: string
          description: Image URL or content in base64
        type:
          description: Specify the image source type.
          type: string
          enum:
            - url
            - base64
      xml:
        name: Image
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        type:
          type: string